fix: move configs
